policy penelitian done

+++ app/Http/Controllers/PenelitianController.php
@@ -9,10 +9,9 @@ namespace App\Http\Controllers;
 
 use App\Exports\PenelitiansExport;
 use App\Models\Penelitian;
-use App\Models\Periode;
-use App\Models\Status;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Maatwebsite\Excel\Facades\Excel;
 
@@ -20,16 +19,24 @@ class PenelitianController extends Controller
 {
     public function index()
     {
-        $data    = Penelitian::all();
-        $periode = Periode::join('penelitians', 'periodes.id', '=', 'penelitians.periode_id')
-            ->where('semester', 1)
-            ->get();
-        return view('penelitian.index', compact('data', 'periode'));
+        $user = Auth::user()->username;
+        if (Auth::user()->role_id !== 3) {
+            $data = Penelitian::orderBy('created_at')->get();
+        } else {
+            $dosen = DB::table('dosens')
+                ->where('user_id', '=', $user)
+                ->get();
+            $data = Penelitian::where('dosen_id', $dosen[0]->id)->get();
+        }
+        return view('penelitian.index', compact('data'));
     }
 
     public function create()
     {
+        $this->authorize('create', Penelitian::class);
+
         $dosen = User::join('dosens', 'users.username', '=', 'dosens.user_id')
+            ->where('dosens.status', 'aktif')
             ->orderBy('dosens.created_at', 'desc')
             ->get();
         $periode = DB::table('periodes')
@@ -43,12 +50,16 @@ class PenelitianController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create', Penelitian::class);
+
         $this->validate($request, [
+            'dosen_id'          => 'required',
             'judul_penelitian'  => 'required',
             'jumlah_anggota'    => 'required',
         ]);
 
         $data = Penelitian::create([
+            'dosen_id'          => $request->dosen_id,
             'status_id'         => $request->status_id,
             'periode_id'        => $request->periode_id,
             'judul_penelitian'  => $request->judul_penelitian,
@@ -62,17 +73,24 @@ class PenelitianController extends Controller
 
     public function edit(Penelitian $penelitian)
     {
+        $this->authorize('update', Penelitian::class);
+
         $periode = DB::table('periodes')
             ->where('semester', '=', 1)
             ->get();
         $status = DB::table('statuses')
             ->where('group', '=', 'penelitian')
             ->get();
-        return view('penelitian.edit', compact('penelitian', 'periode', 'status'));
+        $dosen = DB::table('dosens')
+            ->where('status', '=', 'aktif')
+            ->get();
+        return view('penelitian.edit', compact('penelitian', 'periode', 'status', 'dosen'));
     }
 
     public function update(Request $request, Penelitian $penelitian)
     {
+        $this->authorize('update', Penelitian::class);
+
         $penelitian = Penelitian::findOrFail($penelitian->id);
 
         $penelitian->update([
@@ -89,6 +107,8 @@ class PenelitianController extends Controller
 
     public function destroy(Penelitian $penelitian)
     {
+        $this->authorize('delete', Penelitian::class);
+
         $penelitian->find($penelitian->id)->all();
 
         $penelitian->delete();
